+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.1] - 2018-08-29
+### Changed
+- The promise resolve for the `update()` method now returns the requested data objects
+
+### Fixed
+- Fixed a bug where the `save()` method would not return the promise
+
 ## [0.4.0] - 2018-08-29
 ### Added
 - Added the following options
@@ -81,6 +88,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release
 
+[0.4.1]: https://www.npmjs.com/package/poe-ninja-api-manager/v/0.4.1
 [0.4.0]: https://www.npmjs.com/package/poe-ninja-api-manager/v/0.4.0
 [0.3.3]: https://www.npmjs.com/package/poe-ninja-api-manager/v/0.3.3
 [0.3.2]: https://www.npmjs.com/package/poe-ninja-api-manager/v/0.3.2